fix loading spinner
The loading spinner is now correctly centered in every page.

The loading spinner used to be centered, but by using "position:
absolute". The commit 4c82d73f removed that behaviour because it didn't
look good, in favor of non-centered spinners.
This commit finally addresses the centering of loading spinner without
using "position: absolute".