remove article content from list
It's pointless to send the whole article's content from the endpoints
which return a list of articles. The article's content is something
strictly related to the article details.

By sending the whole article's content from the endpoints which return a
list of articles, the response payload becomes much bigger in size. That
slowdowns loading times, and it's also a waste of bandwidth.